api1/llm/requirements.txt
transformers>=4.45.1
python-dotenv>=1.0.1
click>=8.1.7
flask>=3.0.3
waitress>=3.0.0
torch>=2.3.1
accelerate>=0.25.0 